내용 : 컨트롤러에서 서비스로 비즈니스 로직이동 컨트롤러가 HTTP 요청 처리만을 담당하도록 비즈니스 로직을 서비스 계층에서 처리하도록 처리 개발자 : 양지안

+++ src/main/java/community/whatever/onembackendjava/url/UrlShortenServiceImpl.java
@@ -28,13 +28,30 @@ public class UrlShortenServiceImpl implements  UrlShortenService{
     }
 
     public String createKey(String url) {
-        return urlShortenRepository.createKey(url);
+        String result = "";
+        if(urlShortenRepository.existKey(url)){
+            result = urlShortenRepository.searchKey(url);
+        }else{
+            result = urlShortenRepository.createKey(url);
+        }
+        System.out.println("result = " + result);
+        return result;
     }
 
     public String searchUrl(String key) {
-        return urlShortenRepository.searchUrl(key);
+        if(urlShortenRepository.existKey(key)){
+            return urlShortenRepository.searchUrl(key);
+        }else{
+            // 키가 존재 하지 않으면 Exception
+            throw new IllegalArgumentException("Invalid key");
+        }
     }
-    public int deleteKey(String key) {
-        return urlShortenRepository.deleteKey(key);
+    public boolean deleteKey(String key) {
+        if(urlShortenRepository.existKey(key)){
+            return urlShortenRepository.deleteKey(key);
+        }else{
+            throw new IllegalArgumentException("Invalid key");
+        }
+
     }
 }
